+++ attribution/src/main/java/com/affise/attribution/usecase/DiskUseCase.kt
@@ -1,0 +1,8 @@
+package com.affise.attribution.usecase
+
+interface DiskUseCase {
+
+    fun getTotalDisk(): Long
+
+    fun getFreeDisk(): Long
+}