+++ attribution/src/main/java/com/affise/attribution/usecase/ProcessInfoUseCase.kt
@@ -1,0 +1,6 @@
+package com.affise.attribution.usecase
+
+interface ProcessInfoUseCase {
+
+    fun getCpuCores(): Long
+}