+++ attribution/src/main/java/com/affise/attribution/usecase/ProcessInfoUseCaseImpl.kt
@@ -1,0 +1,6 @@
+package com.affise.attribution.usecase
+
+internal class ProcessInfoUseCaseImpl : ProcessInfoUseCase {
+
+    override fun getCpuCores(): Long = Runtime.getRuntime().availableProcessors().toLong()
+}