+++ attribution/src/main/java/com/affise/attribution/usecase/ScreenUseCase.kt
@@ -1,0 +1,10 @@
+package com.affise.attribution.usecase
+
+interface ScreenUseCase {
+
+    fun getScreenWidth(): Long
+
+    fun getScreenHeight(): Long
+
+    fun getDensity(): Float
+}